+++ posce/comms/edit.py
@@ -14,34 +14,18 @@ from posce            import tools
     default = '',
     type    = str,
 )
-@click.option('-n', '--new',
-    help    = 'Create new note.',
-    is_flag = True,
-)
 @click.pass_obj
-def edit(book, name, editor, new):
+def edit(book, name, editor):
     '''
     Edit a note.
     '''
 
-    notes = list(book.disambiguate(name))
-    efunc = lambda note: click.edit(
+    note = tools.clui.disambiguate(book, name)
+    func = lambda note: click.edit(
         editor    = editor or None,
         extension = f'.{note.ext}',
         filename  = note.path,
     )
 
-    if len(notes) == 0:
-        if new:
-            note = book.create(name, '')
-            if data := efunc(note):
-                note.write(data)
-        else:
-            tools.clui.error(f'Note {name!r} does not exist.')
-
-    elif len(notes) == 1:
-        if data := efunc(notes[0]):
-            notes[0].write(data)
-
-    else:
-        tools.clui.disambiguate(book, name)
+    if data := func(note):
+        note.write(data)