+++ tests/test_comms/test_edit.py
@@ -12,9 +12,8 @@ def test_edit(monkeypatch, book):
     # setup
     editargs = {}
     monkeypatch.setattr(click, 'edit', lambda **k: editargs.update(k))
-    book.create('bravo2', '')
 
-    # success - existing note
+    # success - defaults
     assert out(book, edit, 'alpha') == []
     assert editargs == {
         'editor':    None,
@@ -22,23 +21,10 @@ def test_edit(monkeypatch, book):
         'filename':  book['alpha'].path,
     }
 
-    # success - existing note, custom editor
+    # success - custom editor
     assert out(book, edit, 'alpha', '-e', 'test') == []
     assert editargs == {
         'editor':    'test',
         'extension': '.txt',
         'filename':  book['alpha'].path,
     }
-
-    # success - new note with --new
-    assert out(book, edit, 'new1', '-n') == []
-    assert editargs == {
-        'editor':    None,
-        'extension': '.txt',
-        'filename':  book['new1'].path,
-    }
-
-    # failure - new note without --new
-    assert out(book, edit, 'new2') == [
-        "Error: Note 'new2' does not exist.\n",
-    ]